+++ cTraining/cTraining/cTraining.vcxproj.filters
@@ -15,7 +15,7 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="cTraining.c">
+    <ClCompile Include="CalcCoins.c">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
